docker: build utility inside "local" container
Instead of relying on a utility package to be already build, perform the
build inside the Docker container. This copies the package contents
over, triggers a build, readies a `.whl` file and cleans up. The
"external" stage now downloads the `.whl` file from pip instead of
directly installing it. Last, the final stage will be responsible for
installing the package in the container's virtual environment.

Signed-off-by: James Knight <james.d.knight@live.com>