tox: cleanup linter requirements
The `ruff` package was added to the global development requirements.
However, this is only needed for the tox lint environment. Also, the
explicit dependency on ruff inside the tox configuration defines an
incorrect name of the environment to map the dependency on.

This commit corrects all of the above, in addition to bumping ruff to
its most recent version.

Signed-off-by: James Knight <james.d.knight@live.com>